requirements.txt
asgiref==3.2.10
Django==3.1.1
django-dotenv==1.4.2
gitdb==4.0.5
GitPython==3.1.9
mysqlclient==2.0.1
Pillow==5.4.1
python-decouple==3.3
pytz==2020.1
smmap==3.0.4
sqlparse==0.4.1
